Extract common optimizer tests
